backend/generate-pdf/requirements.txt
psycopg2-binary>=2.9.9
reportlab>=4.0.0
PyPDF2>=3.0.0
